Add WaveFile.Write & Fix Dx7Voice.Active
Also rewite File.StreamHelper

The Dx7Voice.Active bug will cause the Mix to allocate more and more
	voices since none of the Dx7Voice will is not Active.

Use bit shifts to do the stream read and write in StreamHelper.

+++ File/Midi/MidiFile.cs
@@ -36,7 +36,7 @@ namespace Midif {
 			public Chunk (Stream stream) {
 				Id = StreamHelperBe.ReadString(stream, 4);
 
-				Length = StreamHelperBe.ReadUInt32(stream);
+				Length = (int)StreamHelperBe.ReadUInt32(stream);
 //				DebugConsole.WriteLine("Padding " + (Length & 1));
 				Bytes = new byte[Length];
 				stream.Read(Bytes, 0, Length);
